# spotly-backend/openapi.json
{"openapi":"3.0.0","paths":{"/api/parking-spots":{"post":{"description":"Creates a new parking spot with the provided details","operationId":"ParkingSpotsController_create","parameters":[],"requestBody":{"required":true,"description":"Parking spot creation data","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateParkingSpotDto"}}}},"responses":{"201":{"description":"Parking spot successfully created","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ParkingSpotResponseDto"}}}},"400":{"description":"Bad request - Invalid input data"}},"summary":"Create a new parking spot","tags":["parking-spots"]},"get":{"description":"Retrieves a list of all parking spots","operationId":"ParkingSpotsController_listParkingSpots","parameters":[],"responses":{"200":{"description":"Parking spots successfully retrieved","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ParkingSpotsListResponseDto"}}}}},"summary":"List all parking spots","tags":["parking-spots"]}},"/api/parking-spots/{id}":{"delete":{"description":"Deletes a parking spot by its ID","operationId":"ParkingSpotsController_deleteParkingSpot","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"204":{"description":"Parking spot successfully deleted"}},"summary":"Delete a parking spot","tags":["parking-spots"]}},"/api/users":{"post":{"description":"Creates a new user with the provided details","operationId":"UsersController_createUser","parameters":[],"requestBody":{"required":true,"description":"User creation data","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateUserDto"}}}},"responses":{"201":{"description":"User successfully created","content":{"application/json":{"schema":{"type":"object"}}}},"400":{"description":"Bad request - Invalid input data"}},"summary":"Create a new user","tags":["users"]}},"/api/settings":{"post":{"description":"Creates new settings for the specified user","operationId":"SettingsController_createSettings","parameters":[{"name":"userId","required":true,"in":"query","description":"User ID","schema":{"type":"string"}}],"requestBody":{"required":true,"description":"Settings creation data","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateSettingsDto"}}}},"responses":{"201":{"description":"Settings successfully created","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SettingsDto"}}}},"400":{"description":"Bad request - Invalid input data"}},"summary":"Create settings for a user","tags":["settings"]}},"/api/settings/user/{userId}":{"get":{"description":"Retrieves settings for the specified user","operationId":"SettingsController_getSettingsByUserId","parameters":[{"name":"userId","required":true,"in":"path","description":"User ID","schema":{"type":"string"}}],"responses":{"200":{"description":"Settings successfully retrieved","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SettingsDto"}}}},"404":{"description":"Settings not found for the user"}},"summary":"Get settings by user ID","tags":["settings"]}}},"info":{"title":"Spotly Backend API","description":"API documentation for Spotly Backend","version":"1.0","contact":{}},"tags":[{"name":"parking-spots","description":"Parking spots management"},{"name":"users","description":"User management"},{"name":"settings","description":"User settings management"}],"servers":[],"components":{"schemas":{"CreateParkingSpotDto":{"type":"object","properties":{"name":{"type":"string","minLength":1},"description":{"type":"string"},"latitude":{"type":"number","minimum":-90,"maximum":90},"longitude":{"type":"number","minimum":-180,"maximum":180},"paid":{"default":false,"type":"boolean"}},"required":["name","latitude","longitude"]},"ParkingSpotResponseDto":{"type":"object","properties":{"id":{"type":"string","format":"uuid","pattern":"^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$","x-nestjs_zod-has-null":true},"name":{"type":"string","x-nestjs_zod-has-null":true},"description":{"anyOf":[{"type":"string"},{"type":"null"}],"x-nestjs_zod-has-null":true,"x-nestjs_zod-empty-type":true},"latitude":{"type":"number","x-nestjs_zod-has-null":true},"longitude":{"type":"number","x-nestjs_zod-has-null":true},"createdAt":{"type":"string","pattern":"^\\d{4}-\\d{2}-\\d{2}T\\d{2}:\\d{2}:\\d{2}(\\.\\d{3})?Z?$","x-nestjs_zod-has-null":true},"updatedAt":{"type":"string","pattern":"^\\d{4}-\\d{2}-\\d{2}T\\d{2}:\\d{2}:\\d{2}(\\.\\d{3})?Z?$","x-nestjs_zod-has-null":true},"paid":{"type":"boolean","x-nestjs_zod-has-null":true}},"required":["id","name","description","latitude","longitude","createdAt","updatedAt","paid"]},"ParkingSpotsListResponseDto":{"type":"object","properties":{"parkingSpots":{"type":"array","items":{"type":"object","properties":{"id":{"type":"string","format":"uuid","pattern":"^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"},"name":{"type":"string"},"description":{"anyOf":[{"type":"string"},{"type":"null"}]},"latitude":{"type":"number"},"longitude":{"type":"number"},"createdAt":{"type":"string","pattern":"^\\d{4}-\\d{2}-\\d{2}T\\d{2}:\\d{2}:\\d{2}(\\.\\d{3})?Z?$"},"updatedAt":{"type":"string","pattern":"^\\d{4}-\\d{2}-\\d{2}T\\d{2}:\\d{2}:\\d{2}(\\.\\d{3})?Z?$"},"paid":{"type":"boolean"}},"required":["id","name","description","latitude","longitude","createdAt","updatedAt","paid"]},"x-nestjs_zod-has-null":true}},"required":["parkingSpots"]},"CreateUserDto":{"type":"object","properties":{"name":{"type":"string","minLength":1},"surname":{"type":"string","minLength":1},"phone":{"type":"string"},"email":{"type":"string","format":"email","pattern":"^(?!\\.)(?!.*\\.\\.)([A-Za-z0-9_'+\\-\\.]*)[A-Za-z0-9_+-]@([A-Za-z0-9][A-Za-z0-9\\-]*\\.)+[A-Za-z]{2,}$"}},"required":["name","surname","email"]},"CreateSettingsDto":{"type":"object","properties":{"notificationsRadius":{"type":"number"},"showPaidSpots":{"type":"boolean"},"showUnpaidSpots":{"type":"boolean"}},"required":["notificationsRadius","showPaidSpots","showUnpaidSpots"]},"SettingsDto":{"type":"object","properties":{"id":{"type":"string","format":"uuid","pattern":"^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"},"createdAt":{"type":"string","format":"date-time","pattern":"^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?(?:Z))$"},"updatedAt":{"type":"string","format":"date-time","pattern":"^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?(?:Z))$"},"notificationsRadius":{"type":"number"},"showPaidSpots":{"type":"boolean"},"showUnpaidSpots":{"type":"boolean"},"userId":{"type":"string","format":"uuid","pattern":"^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"}},"required":["id","createdAt","updatedAt","notificationsRadius","showPaidSpots","showUnpaidSpots","userId"]}}}}